move getCookie
