+++ api/src/api/v1/__tests__/auth_flow.test.integration.ts
@@ -1,7 +1,7 @@
 import * as Hapi from 'hapi';
 import { init } from '../../../server';
 import { getConfig } from '../../../../config';
-import { getCookie } from '../../../utils';
+import { getCookie } from '../../../../tests/utils/server';
 
 
 describe('Auth flow', () => {
