+++ api/src/api/v1/users/__tests__/login.test.integration.ts
@@ -3,7 +3,7 @@ import * as Knex from 'knex';
 import * as JWT from 'jsonwebtoken';
 import { init } from '../../../../server';
 import { getConfig } from '../../../../../config';
-import { getCookie } from '../../../../utils';
+import { getCookie } from '../../../../../tests/utils/server';
 import { getTrx } from '../../../../../tests/utils/database';
